More study request field components
This completes the conversion of study-specific fields; now we need to also do some of the fields that are shared between single-request and bulk-request (step 3).
