+++ web/components/requests/FcCardStudyRequest.vue
@@ -18,28 +18,16 @@
       <div class="mx-9">
         <v-row>
           <v-col class="py-2" cols="6">
-            <FcSelectEnum
-              v-model="v.studyType.$model"
+            <FcStudyRequestStudyType
               dense
               :disabled="!selected"
-              :error-messages="errorMessagesStudyType"
-              hide-details="auto"
-              item-text="label"
-              label="Study Type"
-              :of-type="StudyType"
-              outlined />
+              :v="v" />
           </v-col>
           <v-col class="py-2" cols="6">
-            <v-select
-              v-model="internalDaysOfWeek"
+            <FcStudyRequestDaysOfWeek
               dense
               :disabled="!selected"
-              :error-messages="errorMessagesDaysOfWeek"
-              hide-details="auto"
-              :items="itemsDaysOfWeek"
-              label="Days"
-              multiple
-              outlined />
+              :v="v" />
           </v-col>
         </v-row>
         <v-row>
@@ -58,16 +46,9 @@
         </v-row>
         <v-row>
           <v-col class="py-2" cols="12">
-            <v-textarea
-              v-model="v.notes.$model"
+            <FcStudyRequestNotes
               :disabled="!selected"
-              :error-messages="errorMessagesNotes"
-              label="Additional Information"
-              :messages="messagesNotes"
-              no-resize
-              outlined
-              rows="4"
-              @blur="v.notes.$touch()"></v-textarea>
+              :v="v" />
           </v-col>
         </v-row>
       </div>
@@ -76,28 +57,25 @@
 </template>
 
 <script>
-import ArrayUtils from '@/lib/ArrayUtils';
 import { StudyHours, StudyType } from '@/lib/Constants';
-import {
-  OPTIONAL,
-  REQUEST_STUDY_REQUIRES_DAYS_OF_WEEK,
-  REQUEST_STUDY_OTHER_HOURS_REQUIRES_NOTES,
-  REQUEST_STUDY_REQUIRES_STUDY_TYPE,
-} from '@/lib/i18n/Strings';
-import TimeFormatters from '@/lib/time/TimeFormatters';
 import FcTextMostRecent from '@/web/components/data/FcTextMostRecent.vue';
-import FcSelectEnum from '@/web/components/inputs/FcSelectEnum.vue';
 import FcIconLocationMulti from '@/web/components/location/FcIconLocationMulti.vue';
+import FcStudyRequestDaysOfWeek
+  from '@/web/components/requests/fields/FcStudyRequestDaysOfWeek.vue';
 import FcStudyRequestDuration from '@/web/components/requests/fields/FcStudyRequestDuration.vue';
 import FcStudyRequestHours from '@/web/components/requests/fields/FcStudyRequestHours.vue';
+import FcStudyRequestNotes from '@/web/components/requests/fields/FcStudyRequestNotes.vue';
+import FcStudyRequestStudyType from '@/web/components/requests/fields/FcStudyRequestStudyType.vue';
 
 export default {
   name: 'FcCardStudyRequest',
   components: {
     FcIconLocationMulti,
-    FcSelectEnum,
+    FcStudyRequestDaysOfWeek,
     FcStudyRequestDuration,
     FcStudyRequestHours,
+    FcStudyRequestNotes,
+    FcStudyRequestStudyType,
     FcTextMostRecent,
   },
   props: {
@@ -113,73 +91,6 @@ export default {
       StudyHours,
       StudyType,
     };
-  },
-  computed: {
-    errorMessagesDaysOfWeek() {
-      const errors = [];
-      if (!this.v.daysOfWeek.required) {
-        errors.push(REQUEST_STUDY_REQUIRES_DAYS_OF_WEEK.text);
-      }
-      if (!this.v.daysOfWeek.$dirty && !this.v.duration.$dirty) {
-        return errors;
-      }
-      const { duration } = this.internalValue;
-      if (!this.v.duration.needsValidDaysOfWeek) {
-        const n = duration / 24;
-        const msg = `Please select ${n} consecutive days or reduce study duration.`;
-        errors.push(msg);
-      }
-      return errors;
-    },
-    errorMessagesNotes() {
-      const errors = [];
-      if (!this.v.notes.requiredIfOtherHours) {
-        errors.push(REQUEST_STUDY_OTHER_HOURS_REQUIRES_NOTES.text);
-      }
-      return errors;
-    },
-    errorMessagesStudyType() {
-      const errors = [];
-      if (!this.v.studyType.required) {
-        errors.push(REQUEST_STUDY_REQUIRES_STUDY_TYPE.text);
-      }
-      return errors;
-    },
-    internalDaysOfWeek: {
-      get() {
-        return this.studyRequest.daysOfWeek;
-      },
-      set(daysOfWeek) {
-        this.studyRequest.daysOfWeek = ArrayUtils.sortBy(daysOfWeek, i => i);
-      },
-    },
-    itemsDaysOfWeek() {
-      return TimeFormatters.DAYS_OF_WEEK.map((text, value) => ({ text, value }));
-    },
-    itemsDuration() {
-      return [
-        { text: '1 day', value: 24 },
-        { text: '2 days', value: 48 },
-        { text: '3 days', value: 72 },
-        { text: '4 days', value: 96 },
-        { text: '5 days', value: 120 },
-        { text: '1 week', value: 168 },
-      ];
-    },
-    messagesDuration() {
-      const { duration } = this.studyRequest;
-      if (duration === null) {
-        return [];
-      }
-      return [`${duration} hours`];
-    },
-    messagesNotes() {
-      const { hours } = this.studyRequest;
-      if (hours === StudyHours.OTHER) {
-        return [];
-      }
-      return [OPTIONAL.text];
-    },
   },
 };
 </script>
